requirements.txt
amqp==2.4.2
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-cors-headers==2.5.2
django-extra-fields==1.1.0
django-filter==2.1.0
django-import-export==1.2.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-redis==4.10.0
djangorestframework==3.9.2
djangorestframework-camel-case==1.0.3
djangorestframework-simplejwt==4.1.0
drf-writable-nested==0.5.1
drf-yasg==1.14.0
et-xmlfile==1.0.1
future==0.17.1
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kombu==4.5.0
liqpay-python==1.0
MarkupSafe==1.1.1
mypy==0.670
mypy-extensions==0.4.1
odfpy==1.4.0
openpyxl==2.6.2
pdfkit==0.6.1
Pillow==5.4.1
psycopg2-binary==2.7.7
PyJWT==1.7.1
python-http-client==3.1.0
pytz==2018.9
PyYAML==5.1
redis==3.2.1
requests==2.21.0
requests-file==1.4.3
ruamel.yaml==0.15.89
sendgrid==6.0.4
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
tldextract==2.2.1
typed-ast==1.3.1
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.24.1
viberbot==1.0.11
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
